+++ cli/src/commands/describe.rs
@@ -29,10 +29,12 @@ use crate::cli_util::RevisionArg;
 use crate::command_error::user_error;
 use crate::command_error::CommandError;
 use crate::complete;
+use crate::description_util::add_trailers_with_template;
 use crate::description_util::description_template;
 use crate::description_util::edit_description;
 use crate::description_util::edit_multiple_descriptions;
 use crate::description_util::join_message_paragraphs;
+use crate::description_util::parse_trailers_template;
 use crate::description_util::ParsedBulkEditMessage;
 use crate::text_util::parse_author;
 use crate::ui::Ui;
@@ -180,6 +182,20 @@ pub(crate) fn cmd_describe(
             commit_builder
         })
         .collect_vec();
+
+    if let Some(trailer_template) = parse_trailers_template(ui, &tx)? {
+        for commit_builder in &mut commit_builders {
+            // The first trailer would become the first line of the description.
+            // Also, a commit with no description is treated in a special way in jujutsu: it
+            // can be discarded as soon as it's no longer the working copy. Adding a
+            // trailer to an empty description would break that logic.
+            if use_editor || !commit_builder.description().is_empty() {
+                let temp_commit = commit_builder.write_hidden()?;
+                let new_description = add_trailers_with_template(&trailer_template, &temp_commit)?;
+                commit_builder.set_description(new_description);
+            }
+        }
+    }
 
     if use_editor {
         let temp_commits: Vec<_> = iter::zip(&commits, &commit_builders)